@@ -1,7 +1,7 @@
 // Code generated by 'yaegi extract bytes'. DO NOT EDIT.
 
-//go:build go1.17
-// +build go1.17
+//go:build go1.19
+// +build go1.19
 
 package sgolib
 
@@ -20,6 +20,7 @@ func init() {
 		"ContainsAny":     reflect.ValueOf(bytes.ContainsAny),
 		"ContainsRune":    reflect.ValueOf(bytes.ContainsRune),
 		"Count":           reflect.ValueOf(bytes.Count),
+		"Cut":             reflect.ValueOf(bytes.Cut),
 		"Equal":           reflect.ValueOf(bytes.Equal),
 		"EqualFold":       reflect.ValueOf(bytes.EqualFold),
 		"ErrTooLarge":     reflect.ValueOf(&bytes.ErrTooLarge).Elem(),
