@@ -1,7 +1,7 @@
 // Code generated by 'yaegi extract context'. DO NOT EDIT.
 
-//go:build go1.17
-// +build go1.17
+//go:build go1.19
+// +build go1.19
 
 package sgolib
 
@@ -38,10 +38,18 @@ type _context_Context struct {
 	WDeadline func() (deadline time.Time, ok bool)
 	WDone     func() <-chan struct{}
 	WErr      func() error
-	WValue    func(key interface{}) interface{}
+	WValue    func(key any) any
 }
 
-func (W _context_Context) Deadline() (deadline time.Time, ok bool) { return W.WDeadline() }
-func (W _context_Context) Done() <-chan struct{}                   { return W.WDone() }
-func (W _context_Context) Err() error                              { return W.WErr() }
-func (W _context_Context) Value(key interface{}) interface{}       { return W.WValue(key) }
+func (W _context_Context) Deadline() (deadline time.Time, ok bool) {
+	return W.WDeadline()
+}
+func (W _context_Context) Done() <-chan struct{} {
+	return W.WDone()
+}
+func (W _context_Context) Err() error {
+	return W.WErr()
+}
+func (W _context_Context) Value(key any) any {
+	return W.WValue(key)
+}
