@@ -1,7 +1,7 @@
 // Code generated by 'yaegi extract encoding/json'. DO NOT EDIT.
 
-//go:build go1.17
-// +build go1.17
+//go:build go1.19
+// +build go1.19
 
 package sgolib
 
@@ -54,7 +54,9 @@ type _encoding_json_Marshaler struct {
 	WMarshalJSON func() ([]byte, error)
 }
 
-func (W _encoding_json_Marshaler) MarshalJSON() ([]byte, error) { return W.WMarshalJSON() }
+func (W _encoding_json_Marshaler) MarshalJSON() ([]byte, error) {
+	return W.WMarshalJSON()
+}
 
 // _encoding_json_Token is an interface wrapper for Token type
 type _encoding_json_Token struct {
@@ -67,4 +69,6 @@ type _encoding_json_Unmarshaler struct {
 	WUnmarshalJSON func(a0 []byte) error
 }
 
-func (W _encoding_json_Unmarshaler) UnmarshalJSON(a0 []byte) error { return W.WUnmarshalJSON(a0) }
+func (W _encoding_json_Unmarshaler) UnmarshalJSON(a0 []byte) error {
+	return W.WUnmarshalJSON(a0)
+}
