@@ -1,7 +1,7 @@
 // Code generated by 'yaegi extract flag'. DO NOT EDIT.
 
-//go:build go1.17
-// +build go1.17
+//go:build go1.19
+// +build go1.19
 
 package sgolib
 
@@ -41,6 +41,7 @@ func init() {
 		"Set":             reflect.ValueOf(flag.Set),
 		"String":          reflect.ValueOf(flag.String),
 		"StringVar":       reflect.ValueOf(flag.StringVar),
+		"TextVar":         reflect.ValueOf(flag.TextVar),
 		"Uint":            reflect.ValueOf(flag.Uint),
 		"Uint64":          reflect.ValueOf(flag.Uint64),
 		"Uint64Var":       reflect.ValueOf(flag.Uint64Var),
@@ -67,14 +68,23 @@ func init() {
 // _flag_Getter is an interface wrapper for Getter type
 type _flag_Getter struct {
 	IValue  interface{}
-	WGet    func() interface{}
+	WGet    func() any
 	WSet    func(a0 string) error
 	WString func() string
 }
 
-func (W _flag_Getter) Get() interface{}    { return W.WGet() }
-func (W _flag_Getter) Set(a0 string) error { return W.WSet(a0) }
-func (W _flag_Getter) String() string      { return W.WString() }
+func (W _flag_Getter) Get() any {
+	return W.WGet()
+}
+func (W _flag_Getter) Set(a0 string) error {
+	return W.WSet(a0)
+}
+func (W _flag_Getter) String() string {
+	if W.WString == nil {
+		return ""
+	}
+	return W.WString()
+}
 
 // _flag_Value is an interface wrapper for Value type
 type _flag_Value struct {
@@ -83,5 +93,12 @@ type _flag_Value struct {
 	WString func() string
 }
 
-func (W _flag_Value) Set(a0 string) error { return W.WSet(a0) }
-func (W _flag_Value) String() string      { return W.WString() }
+func (W _flag_Value) Set(a0 string) error {
+	return W.WSet(a0)
+}
+func (W _flag_Value) String() string {
+	if W.WString == nil {
+		return ""
+	}
+	return W.WString()
+}
