@@ -1,7 +1,7 @@
 // Code generated by 'yaegi extract fmt'. DO NOT EDIT.
 
-//go:build go1.17
-// +build go1.17
+//go:build go1.19
+// +build go1.19
 
 package sgolib
 
@@ -13,6 +13,9 @@ import (
 func init() {
 	Symbols["fmt/fmt"] = map[string]reflect.Value{
 		// function, constant and variable definitions
+		"Append":   reflect.ValueOf(fmt.Append),
+		"Appendf":  reflect.ValueOf(fmt.Appendf),
+		"Appendln": reflect.ValueOf(fmt.Appendln),
 		"Errorf":   reflect.ValueOf(fmt.Errorf),
 		"Fprint":   reflect.ValueOf(fmt.Fprint),
 		"Fprintf":  reflect.ValueOf(fmt.Fprintf),
@@ -57,7 +60,9 @@ type _fmt_Formatter struct {
 	WFormat func(f fmt.State, verb rune)
 }
 
-func (W _fmt_Formatter) Format(f fmt.State, verb rune) { W.WFormat(f, verb) }
+func (W _fmt_Formatter) Format(f fmt.State, verb rune) {
+	W.WFormat(f, verb)
+}
 
 // _fmt_GoStringer is an interface wrapper for GoStringer type
 type _fmt_GoStringer struct {
@@ -65,7 +70,9 @@ type _fmt_GoStringer struct {
 	WGoString func() string
 }
 
-func (W _fmt_GoStringer) GoString() string { return W.WGoString() }
+func (W _fmt_GoStringer) GoString() string {
+	return W.WGoString()
+}
 
 // _fmt_ScanState is an interface wrapper for ScanState type
 type _fmt_ScanState struct {
@@ -78,14 +85,24 @@ type _fmt_ScanState struct {
 	WWidth      func() (wid int, ok bool)
 }
 
-func (W _fmt_ScanState) Read(buf []byte) (n int, err error)      { return W.WRead(buf) }
-func (W _fmt_ScanState) ReadRune() (r rune, size int, err error) { return W.WReadRune() }
-func (W _fmt_ScanState) SkipSpace()                              { W.WSkipSpace() }
+func (W _fmt_ScanState) Read(buf []byte) (n int, err error) {
+	return W.WRead(buf)
+}
+func (W _fmt_ScanState) ReadRune() (r rune, size int, err error) {
+	return W.WReadRune()
+}
+func (W _fmt_ScanState) SkipSpace() {
+	W.WSkipSpace()
+}
 func (W _fmt_ScanState) Token(skipSpace bool, f func(rune) bool) (token []byte, err error) {
 	return W.WToken(skipSpace, f)
 }
-func (W _fmt_ScanState) UnreadRune() error         { return W.WUnreadRune() }
-func (W _fmt_ScanState) Width() (wid int, ok bool) { return W.WWidth() }
+func (W _fmt_ScanState) UnreadRune() error {
+	return W.WUnreadRune()
+}
+func (W _fmt_ScanState) Width() (wid int, ok bool) {
+	return W.WWidth()
+}
 
 // _fmt_Scanner is an interface wrapper for Scanner type
 type _fmt_Scanner struct {
@@ -93,7 +110,9 @@ type _fmt_Scanner struct {
 	WScan  func(state fmt.ScanState, verb rune) error
 }
 
-func (W _fmt_Scanner) Scan(state fmt.ScanState, verb rune) error { return W.WScan(state, verb) }
+func (W _fmt_Scanner) Scan(state fmt.ScanState, verb rune) error {
+	return W.WScan(state, verb)
+}
 
 // _fmt_State is an interface wrapper for State type
 type _fmt_State struct {
@@ -104,10 +123,18 @@ type _fmt_State struct {
 	WWrite     func(b []byte) (n int, err error)
 }
 
-func (W _fmt_State) Flag(c int) bool                   { return W.WFlag(c) }
-func (W _fmt_State) Precision() (prec int, ok bool)    { return W.WPrecision() }
-func (W _fmt_State) Width() (wid int, ok bool)         { return W.WWidth() }
-func (W _fmt_State) Write(b []byte) (n int, err error) { return W.WWrite(b) }
+func (W _fmt_State) Flag(c int) bool {
+	return W.WFlag(c)
+}
+func (W _fmt_State) Precision() (prec int, ok bool) {
+	return W.WPrecision()
+}
+func (W _fmt_State) Width() (wid int, ok bool) {
+	return W.WWidth()
+}
+func (W _fmt_State) Write(b []byte) (n int, err error) {
+	return W.WWrite(b)
+}
 
 // _fmt_Stringer is an interface wrapper for Stringer type
 type _fmt_Stringer struct {
@@ -115,4 +142,9 @@ type _fmt_Stringer struct {
 	WString func() string
 }
 
-func (W _fmt_Stringer) String() string { return W.WString() }
+func (W _fmt_Stringer) String() string {
+	if W.WString == nil {
+		return ""
+	}
+	return W.WString()
+}
