@@ -1,7 +1,7 @@
 // Code generated by 'yaegi extract io/fs'. DO NOT EDIT.
 
-//go:build go1.17
-// +build go1.17
+//go:build go1.19
+// +build go1.19
 
 package sgolib
 
@@ -82,10 +82,18 @@ type _io_fs_DirEntry struct {
 	WType  func() fs.FileMode
 }
 
-func (W _io_fs_DirEntry) Info() (fs.FileInfo, error) { return W.WInfo() }
-func (W _io_fs_DirEntry) IsDir() bool                { return W.WIsDir() }
-func (W _io_fs_DirEntry) Name() string               { return W.WName() }
-func (W _io_fs_DirEntry) Type() fs.FileMode          { return W.WType() }
+func (W _io_fs_DirEntry) Info() (fs.FileInfo, error) {
+	return W.WInfo()
+}
+func (W _io_fs_DirEntry) IsDir() bool {
+	return W.WIsDir()
+}
+func (W _io_fs_DirEntry) Name() string {
+	return W.WName()
+}
+func (W _io_fs_DirEntry) Type() fs.FileMode {
+	return W.WType()
+}
 
 // _io_fs_FS is an interface wrapper for FS type
 type _io_fs_FS struct {
@@ -93,7 +101,9 @@ type _io_fs_FS struct {
 	WOpen  func(name string) (fs.File, error)
 }
 
-func (W _io_fs_FS) Open(name string) (fs.File, error) { return W.WOpen(name) }
+func (W _io_fs_FS) Open(name string) (fs.File, error) {
+	return W.WOpen(name)
+}
 
 // _io_fs_File is an interface wrapper for File type
 type _io_fs_File struct {
@@ -103,9 +113,15 @@ type _io_fs_File struct {
 	WStat  func() (fs.FileInfo, error)
 }
 
-func (W _io_fs_File) Close() error                { return W.WClose() }
-func (W _io_fs_File) Read(a0 []byte) (int, error) { return W.WRead(a0) }
-func (W _io_fs_File) Stat() (fs.FileInfo, error)  { return W.WStat() }
+func (W _io_fs_File) Close() error {
+	return W.WClose()
+}
+func (W _io_fs_File) Read(a0 []byte) (int, error) {
+	return W.WRead(a0)
+}
+func (W _io_fs_File) Stat() (fs.FileInfo, error) {
+	return W.WStat()
+}
 
 // _io_fs_FileInfo is an interface wrapper for FileInfo type
 type _io_fs_FileInfo struct {
@@ -115,15 +131,27 @@ type _io_fs_FileInfo struct {
 	WMode    func() fs.FileMode
 	WName    func() string
 	WSize    func() int64
-	WSys     func() interface{}
+	WSys     func() any
 }
 
-func (W _io_fs_FileInfo) IsDir() bool        { return W.WIsDir() }
-func (W _io_fs_FileInfo) ModTime() time.Time { return W.WModTime() }
-func (W _io_fs_FileInfo) Mode() fs.FileMode  { return W.WMode() }
-func (W _io_fs_FileInfo) Name() string       { return W.WName() }
-func (W _io_fs_FileInfo) Size() int64        { return W.WSize() }
-func (W _io_fs_FileInfo) Sys() interface{}   { return W.WSys() }
+func (W _io_fs_FileInfo) IsDir() bool {
+	return W.WIsDir()
+}
+func (W _io_fs_FileInfo) ModTime() time.Time {
+	return W.WModTime()
+}
+func (W _io_fs_FileInfo) Mode() fs.FileMode {
+	return W.WMode()
+}
+func (W _io_fs_FileInfo) Name() string {
+	return W.WName()
+}
+func (W _io_fs_FileInfo) Size() int64 {
+	return W.WSize()
+}
+func (W _io_fs_FileInfo) Sys() any {
+	return W.WSys()
+}
 
 // _io_fs_GlobFS is an interface wrapper for GlobFS type
 type _io_fs_GlobFS struct {
@@ -132,8 +160,12 @@ type _io_fs_GlobFS struct {
 	WOpen  func(name string) (fs.File, error)
 }
 
-func (W _io_fs_GlobFS) Glob(pattern string) ([]string, error) { return W.WGlob(pattern) }
-func (W _io_fs_GlobFS) Open(name string) (fs.File, error)     { return W.WOpen(name) }
+func (W _io_fs_GlobFS) Glob(pattern string) ([]string, error) {
+	return W.WGlob(pattern)
+}
+func (W _io_fs_GlobFS) Open(name string) (fs.File, error) {
+	return W.WOpen(name)
+}
 
 // _io_fs_ReadDirFS is an interface wrapper for ReadDirFS type
 type _io_fs_ReadDirFS struct {
@@ -142,8 +174,12 @@ type _io_fs_ReadDirFS struct {
 	WReadDir func(name string) ([]fs.DirEntry, error)
 }
 
-func (W _io_fs_ReadDirFS) Open(name string) (fs.File, error)          { return W.WOpen(name) }
-func (W _io_fs_ReadDirFS) ReadDir(name string) ([]fs.DirEntry, error) { return W.WReadDir(name) }
+func (W _io_fs_ReadDirFS) Open(name string) (fs.File, error) {
+	return W.WOpen(name)
+}
+func (W _io_fs_ReadDirFS) ReadDir(name string) ([]fs.DirEntry, error) {
+	return W.WReadDir(name)
+}
 
 // _io_fs_ReadDirFile is an interface wrapper for ReadDirFile type
 type _io_fs_ReadDirFile struct {
@@ -154,10 +190,18 @@ type _io_fs_ReadDirFile struct {
 	WStat    func() (fs.FileInfo, error)
 }
 
-func (W _io_fs_ReadDirFile) Close() error                         { return W.WClose() }
-func (W _io_fs_ReadDirFile) Read(a0 []byte) (int, error)          { return W.WRead(a0) }
-func (W _io_fs_ReadDirFile) ReadDir(n int) ([]fs.DirEntry, error) { return W.WReadDir(n) }
-func (W _io_fs_ReadDirFile) Stat() (fs.FileInfo, error)           { return W.WStat() }
+func (W _io_fs_ReadDirFile) Close() error {
+	return W.WClose()
+}
+func (W _io_fs_ReadDirFile) Read(a0 []byte) (int, error) {
+	return W.WRead(a0)
+}
+func (W _io_fs_ReadDirFile) ReadDir(n int) ([]fs.DirEntry, error) {
+	return W.WReadDir(n)
+}
+func (W _io_fs_ReadDirFile) Stat() (fs.FileInfo, error) {
+	return W.WStat()
+}
 
 // _io_fs_ReadFileFS is an interface wrapper for ReadFileFS type
 type _io_fs_ReadFileFS struct {
@@ -166,8 +210,12 @@ type _io_fs_ReadFileFS struct {
 	WReadFile func(name string) ([]byte, error)
 }
 
-func (W _io_fs_ReadFileFS) Open(name string) (fs.File, error)    { return W.WOpen(name) }
-func (W _io_fs_ReadFileFS) ReadFile(name string) ([]byte, error) { return W.WReadFile(name) }
+func (W _io_fs_ReadFileFS) Open(name string) (fs.File, error) {
+	return W.WOpen(name)
+}
+func (W _io_fs_ReadFileFS) ReadFile(name string) ([]byte, error) {
+	return W.WReadFile(name)
+}
 
 // _io_fs_StatFS is an interface wrapper for StatFS type
 type _io_fs_StatFS struct {
@@ -176,8 +224,12 @@ type _io_fs_StatFS struct {
 	WStat  func(name string) (fs.FileInfo, error)
 }
 
-func (W _io_fs_StatFS) Open(name string) (fs.File, error)     { return W.WOpen(name) }
-func (W _io_fs_StatFS) Stat(name string) (fs.FileInfo, error) { return W.WStat(name) }
+func (W _io_fs_StatFS) Open(name string) (fs.File, error) {
+	return W.WOpen(name)
+}
+func (W _io_fs_StatFS) Stat(name string) (fs.FileInfo, error) {
+	return W.WStat(name)
+}
 
 // _io_fs_SubFS is an interface wrapper for SubFS type
 type _io_fs_SubFS struct {
@@ -186,5 +238,9 @@ type _io_fs_SubFS struct {
 	WSub   func(dir string) (fs.FS, error)
 }
 
-func (W _io_fs_SubFS) Open(name string) (fs.File, error) { return W.WOpen(name) }
-func (W _io_fs_SubFS) Sub(dir string) (fs.FS, error)     { return W.WSub(dir) }
+func (W _io_fs_SubFS) Open(name string) (fs.File, error) {
+	return W.WOpen(name)
+}
+func (W _io_fs_SubFS) Sub(dir string) (fs.FS, error) {
+	return W.WSub(dir)
+}
