@@ -1,7 +1,7 @@
 // Code generated by 'yaegi extract io'. DO NOT EDIT.
 
-//go:build go1.17
-// +build go1.17
+//go:build go1.19
+// +build go1.19
 
 package sgolib
 
@@ -100,7 +100,9 @@ type _io_ByteReader struct {
 	WReadByte func() (byte, error)
 }
 
-func (W _io_ByteReader) ReadByte() (byte, error) { return W.WReadByte() }
+func (W _io_ByteReader) ReadByte() (byte, error) {
+	return W.WReadByte()
+}
 
 // _io_ByteScanner is an interface wrapper for ByteScanner type
 type _io_ByteScanner struct {
@@ -109,8 +111,12 @@ type _io_ByteScanner struct {
 	WUnreadByte func() error
 }
 
-func (W _io_ByteScanner) ReadByte() (byte, error) { return W.WReadByte() }
-func (W _io_ByteScanner) UnreadByte() error       { return W.WUnreadByte() }
+func (W _io_ByteScanner) ReadByte() (byte, error) {
+	return W.WReadByte()
+}
+func (W _io_ByteScanner) UnreadByte() error {
+	return W.WUnreadByte()
+}
 
 // _io_ByteWriter is an interface wrapper for ByteWriter type
 type _io_ByteWriter struct {
@@ -118,7 +124,9 @@ type _io_ByteWriter struct {
 	WWriteByte func(c byte) error
 }
 
-func (W _io_ByteWriter) WriteByte(c byte) error { return W.WWriteByte(c) }
+func (W _io_ByteWriter) WriteByte(c byte) error {
+	return W.WWriteByte(c)
+}
 
 // _io_Closer is an interface wrapper for Closer type
 type _io_Closer struct {
@@ -126,7 +134,9 @@ type _io_Closer struct {
 	WClose func() error
 }
 
-func (W _io_Closer) Close() error { return W.WClose() }
+func (W _io_Closer) Close() error {
+	return W.WClose()
+}
 
 // _io_ReadCloser is an interface wrapper for ReadCloser type
 type _io_ReadCloser struct {
@@ -135,8 +145,12 @@ type _io_ReadCloser struct {
 	WRead  func(p []byte) (n int, err error)
 }
 
-func (W _io_ReadCloser) Close() error                     { return W.WClose() }
-func (W _io_ReadCloser) Read(p []byte) (n int, err error) { return W.WRead(p) }
+func (W _io_ReadCloser) Close() error {
+	return W.WClose()
+}
+func (W _io_ReadCloser) Read(p []byte) (n int, err error) {
+	return W.WRead(p)
+}
 
 // _io_ReadSeekCloser is an interface wrapper for ReadSeekCloser type
 type _io_ReadSeekCloser struct {
@@ -146,8 +160,12 @@ type _io_ReadSeekCloser struct {
 	WSeek  func(offset int64, whence int) (int64, error)
 }
 
-func (W _io_ReadSeekCloser) Close() error                     { return W.WClose() }
-func (W _io_ReadSeekCloser) Read(p []byte) (n int, err error) { return W.WRead(p) }
+func (W _io_ReadSeekCloser) Close() error {
+	return W.WClose()
+}
+func (W _io_ReadSeekCloser) Read(p []byte) (n int, err error) {
+	return W.WRead(p)
+}
 func (W _io_ReadSeekCloser) Seek(offset int64, whence int) (int64, error) {
 	return W.WSeek(offset, whence)
 }
@@ -159,8 +177,12 @@ type _io_ReadSeeker struct {
 	WSeek  func(offset int64, whence int) (int64, error)
 }
 
-func (W _io_ReadSeeker) Read(p []byte) (n int, err error)             { return W.WRead(p) }
-func (W _io_ReadSeeker) Seek(offset int64, whence int) (int64, error) { return W.WSeek(offset, whence) }
+func (W _io_ReadSeeker) Read(p []byte) (n int, err error) {
+	return W.WRead(p)
+}
+func (W _io_ReadSeeker) Seek(offset int64, whence int) (int64, error) {
+	return W.WSeek(offset, whence)
+}
 
 // _io_ReadWriteCloser is an interface wrapper for ReadWriteCloser type
 type _io_ReadWriteCloser struct {
@@ -170,9 +192,15 @@ type _io_ReadWriteCloser struct {
 	WWrite func(p []byte) (n int, err error)
 }
 
-func (W _io_ReadWriteCloser) Close() error                      { return W.WClose() }
-func (W _io_ReadWriteCloser) Read(p []byte) (n int, err error)  { return W.WRead(p) }
-func (W _io_ReadWriteCloser) Write(p []byte) (n int, err error) { return W.WWrite(p) }
+func (W _io_ReadWriteCloser) Close() error {
+	return W.WClose()
+}
+func (W _io_ReadWriteCloser) Read(p []byte) (n int, err error) {
+	return W.WRead(p)
+}
+func (W _io_ReadWriteCloser) Write(p []byte) (n int, err error) {
+	return W.WWrite(p)
+}
 
 // _io_ReadWriteSeeker is an interface wrapper for ReadWriteSeeker type
 type _io_ReadWriteSeeker struct {
@@ -182,11 +210,15 @@ type _io_ReadWriteSeeker struct {
 	WWrite func(p []byte) (n int, err error)
 }
 
-func (W _io_ReadWriteSeeker) Read(p []byte) (n int, err error) { return W.WRead(p) }
+func (W _io_ReadWriteSeeker) Read(p []byte) (n int, err error) {
+	return W.WRead(p)
+}
 func (W _io_ReadWriteSeeker) Seek(offset int64, whence int) (int64, error) {
 	return W.WSeek(offset, whence)
 }
-func (W _io_ReadWriteSeeker) Write(p []byte) (n int, err error) { return W.WWrite(p) }
+func (W _io_ReadWriteSeeker) Write(p []byte) (n int, err error) {
+	return W.WWrite(p)
+}
 
 // _io_ReadWriter is an interface wrapper for ReadWriter type
 type _io_ReadWriter struct {
@@ -195,8 +227,12 @@ type _io_ReadWriter struct {
 	WWrite func(p []byte) (n int, err error)
 }
 
-func (W _io_ReadWriter) Read(p []byte) (n int, err error)  { return W.WRead(p) }
-func (W _io_ReadWriter) Write(p []byte) (n int, err error) { return W.WWrite(p) }
+func (W _io_ReadWriter) Read(p []byte) (n int, err error) {
+	return W.WRead(p)
+}
+func (W _io_ReadWriter) Write(p []byte) (n int, err error) {
+	return W.WWrite(p)
+}
 
 // _io_Reader is an interface wrapper for Reader type
 type _io_Reader struct {
@@ -204,7 +240,9 @@ type _io_Reader struct {
 	WRead  func(p []byte) (n int, err error)
 }
 
-func (W _io_Reader) Read(p []byte) (n int, err error) { return W.WRead(p) }
+func (W _io_Reader) Read(p []byte) (n int, err error) {
+	return W.WRead(p)
+}
 
 // _io_ReaderAt is an interface wrapper for ReaderAt type
 type _io_ReaderAt struct {
@@ -212,7 +250,9 @@ type _io_ReaderAt struct {
 	WReadAt func(p []byte, off int64) (n int, err error)
 }
 
-func (W _io_ReaderAt) ReadAt(p []byte, off int64) (n int, err error) { return W.WReadAt(p, off) }
+func (W _io_ReaderAt) ReadAt(p []byte, off int64) (n int, err error) {
+	return W.WReadAt(p, off)
+}
 
 // _io_ReaderFrom is an interface wrapper for ReaderFrom type
 type _io_ReaderFrom struct {
@@ -220,7 +260,9 @@ type _io_ReaderFrom struct {
 	WReadFrom func(r io.Reader) (n int64, err error)
 }
 
-func (W _io_ReaderFrom) ReadFrom(r io.Reader) (n int64, err error) { return W.WReadFrom(r) }
+func (W _io_ReaderFrom) ReadFrom(r io.Reader) (n int64, err error) {
+	return W.WReadFrom(r)
+}
 
 // _io_RuneReader is an interface wrapper for RuneReader type
 type _io_RuneReader struct {
@@ -228,7 +270,9 @@ type _io_RuneReader struct {
 	WReadRune func() (r rune, size int, err error)
 }
 
-func (W _io_RuneReader) ReadRune() (r rune, size int, err error) { return W.WReadRune() }
+func (W _io_RuneReader) ReadRune() (r rune, size int, err error) {
+	return W.WReadRune()
+}
 
 // _io_RuneScanner is an interface wrapper for RuneScanner type
 type _io_RuneScanner struct {
@@ -237,8 +281,12 @@ type _io_RuneScanner struct {
 	WUnreadRune func() error
 }
 
-func (W _io_RuneScanner) ReadRune() (r rune, size int, err error) { return W.WReadRune() }
-func (W _io_RuneScanner) UnreadRune() error                       { return W.WUnreadRune() }
+func (W _io_RuneScanner) ReadRune() (r rune, size int, err error) {
+	return W.WReadRune()
+}
+func (W _io_RuneScanner) UnreadRune() error {
+	return W.WUnreadRune()
+}
 
 // _io_Seeker is an interface wrapper for Seeker type
 type _io_Seeker struct {
@@ -246,7 +294,9 @@ type _io_Seeker struct {
 	WSeek  func(offset int64, whence int) (int64, error)
 }
 
-func (W _io_Seeker) Seek(offset int64, whence int) (int64, error) { return W.WSeek(offset, whence) }
+func (W _io_Seeker) Seek(offset int64, whence int) (int64, error) {
+	return W.WSeek(offset, whence)
+}
 
 // _io_StringWriter is an interface wrapper for StringWriter type
 type _io_StringWriter struct {
@@ -254,7 +304,9 @@ type _io_StringWriter struct {
 	WWriteString func(s string) (n int, err error)
 }
 
-func (W _io_StringWriter) WriteString(s string) (n int, err error) { return W.WWriteString(s) }
+func (W _io_StringWriter) WriteString(s string) (n int, err error) {
+	return W.WWriteString(s)
+}
 
 // _io_WriteCloser is an interface wrapper for WriteCloser type
 type _io_WriteCloser struct {
@@ -263,8 +315,12 @@ type _io_WriteCloser struct {
 	WWrite func(p []byte) (n int, err error)
 }
 
-func (W _io_WriteCloser) Close() error                      { return W.WClose() }
-func (W _io_WriteCloser) Write(p []byte) (n int, err error) { return W.WWrite(p) }
+func (W _io_WriteCloser) Close() error {
+	return W.WClose()
+}
+func (W _io_WriteCloser) Write(p []byte) (n int, err error) {
+	return W.WWrite(p)
+}
 
 // _io_WriteSeeker is an interface wrapper for WriteSeeker type
 type _io_WriteSeeker struct {
@@ -276,7 +332,9 @@ type _io_WriteSeeker struct {
 func (W _io_WriteSeeker) Seek(offset int64, whence int) (int64, error) {
 	return W.WSeek(offset, whence)
 }
-func (W _io_WriteSeeker) Write(p []byte) (n int, err error) { return W.WWrite(p) }
+func (W _io_WriteSeeker) Write(p []byte) (n int, err error) {
+	return W.WWrite(p)
+}
 
 // _io_Writer is an interface wrapper for Writer type
 type _io_Writer struct {
@@ -284,7 +342,9 @@ type _io_Writer struct {
 	WWrite func(p []byte) (n int, err error)
 }
 
-func (W _io_Writer) Write(p []byte) (n int, err error) { return W.WWrite(p) }
+func (W _io_Writer) Write(p []byte) (n int, err error) {
+	return W.WWrite(p)
+}
 
 // _io_WriterAt is an interface wrapper for WriterAt type
 type _io_WriterAt struct {
@@ -292,7 +352,9 @@ type _io_WriterAt struct {
 	WWriteAt func(p []byte, off int64) (n int, err error)
 }
 
-func (W _io_WriterAt) WriteAt(p []byte, off int64) (n int, err error) { return W.WWriteAt(p, off) }
+func (W _io_WriterAt) WriteAt(p []byte, off int64) (n int, err error) {
+	return W.WWriteAt(p, off)
+}
 
 // _io_WriterTo is an interface wrapper for WriterTo type
 type _io_WriterTo struct {
@@ -300,4 +362,6 @@ type _io_WriterTo struct {
 	WWriteTo func(w io.Writer) (n int64, err error)
 }
 
-func (W _io_WriterTo) WriteTo(w io.Writer) (n int64, err error) { return W.WWriteTo(w) }
+func (W _io_WriterTo) WriteTo(w io.Writer) (n int64, err error) {
+	return W.WWriteTo(w)
+}
