@@ -1,7 +1,7 @@
 // Code generated by 'yaegi extract os/exec'. DO NOT EDIT.
 
-//go:build go1.17
-// +build go1.17
+//go:build go1.19
+// +build go1.19
 
 package sgolib
 
@@ -15,6 +15,7 @@ func init() {
 		// function, constant and variable definitions
 		"Command":        reflect.ValueOf(exec.Command),
 		"CommandContext": reflect.ValueOf(exec.CommandContext),
+		"ErrDot":         reflect.ValueOf(&exec.ErrDot).Elem(),
 		"ErrNotFound":    reflect.ValueOf(&exec.ErrNotFound).Elem(),
 		"LookPath":       reflect.ValueOf(exec.LookPath),
 
