@@ -1,7 +1,7 @@
 // Code generated by 'yaegi extract os/signal'. DO NOT EDIT.
 
-//go:build go1.17
-// +build go1.17
+//go:build go1.19
+// +build go1.19
 
 package sgolib
 
