@@ -1,7 +1,7 @@
 // Code generated by 'yaegi extract runtime'. DO NOT EDIT.
 
-//go:build go1.17
-// +build go1.17
+//go:build go1.19
+// +build go1.19
 
 package sgolib
 
@@ -76,5 +76,9 @@ type _runtime_Error struct {
 	WRuntimeError func()
 }
 
-func (W _runtime_Error) Error() string { return W.WError() }
-func (W _runtime_Error) RuntimeError() { W.WRuntimeError() }
+func (W _runtime_Error) Error() string {
+	return W.WError()
+}
+func (W _runtime_Error) RuntimeError() {
+	W.WRuntimeError()
+}
