@@ -1,7 +1,7 @@
 // Code generated by 'yaegi extract sort'. DO NOT EDIT.
 
-//go:build go1.17
-// +build go1.17
+//go:build go1.19
+// +build go1.19
 
 package sgolib
 
@@ -13,6 +13,7 @@ import (
 func init() {
 	Symbols["sort/sort"] = map[string]reflect.Value{
 		// function, constant and variable definitions
+		"Find":              reflect.ValueOf(sort.Find),
 		"Float64s":          reflect.ValueOf(sort.Float64s),
 		"Float64sAreSorted": reflect.ValueOf(sort.Float64sAreSorted),
 		"Ints":              reflect.ValueOf(sort.Ints),
@@ -50,6 +51,12 @@ type _sort_Interface struct {
 	WSwap  func(i int, j int)
 }
 
-func (W _sort_Interface) Len() int               { return W.WLen() }
-func (W _sort_Interface) Less(i int, j int) bool { return W.WLess(i, j) }
-func (W _sort_Interface) Swap(i int, j int)      { W.WSwap(i, j) }
+func (W _sort_Interface) Len() int {
+	return W.WLen()
+}
+func (W _sort_Interface) Less(i int, j int) bool {
+	return W.WLess(i, j)
+}
+func (W _sort_Interface) Swap(i int, j int) {
+	W.WSwap(i, j)
+}
