@@ -1,7 +1,7 @@
 // Code generated by 'yaegi extract strings'. DO NOT EDIT.
 
-//go:build go1.17
-// +build go1.17
+//go:build go1.19
+// +build go1.19
 
 package sgolib
 
@@ -13,11 +13,13 @@ import (
 func init() {
 	Symbols["strings/strings"] = map[string]reflect.Value{
 		// function, constant and variable definitions
+		"Clone":          reflect.ValueOf(strings.Clone),
 		"Compare":        reflect.ValueOf(strings.Compare),
 		"Contains":       reflect.ValueOf(strings.Contains),
 		"ContainsAny":    reflect.ValueOf(strings.ContainsAny),
 		"ContainsRune":   reflect.ValueOf(strings.ContainsRune),
 		"Count":          reflect.ValueOf(strings.Count),
+		"Cut":            reflect.ValueOf(strings.Cut),
 		"EqualFold":      reflect.ValueOf(strings.EqualFold),
 		"Fields":         reflect.ValueOf(strings.Fields),
 		"FieldsFunc":     reflect.ValueOf(strings.FieldsFunc),
