@@ -1,7 +1,7 @@
 // Code generated by 'yaegi extract sync/atomic'. DO NOT EDIT.
 
-//go:build go1.17
-// +build go1.17
+//go:build go1.19
+// +build go1.19
 
 package sgolib
 
@@ -44,6 +44,12 @@ func init() {
 		"SwapUintptr":           reflect.ValueOf(atomic.SwapUintptr),
 
 		// type definitions
-		"Value": reflect.ValueOf((*atomic.Value)(nil)),
+		"Bool":    reflect.ValueOf((*atomic.Bool)(nil)),
+		"Int32":   reflect.ValueOf((*atomic.Int32)(nil)),
+		"Int64":   reflect.ValueOf((*atomic.Int64)(nil)),
+		"Uint32":  reflect.ValueOf((*atomic.Uint32)(nil)),
+		"Uint64":  reflect.ValueOf((*atomic.Uint64)(nil)),
+		"Uintptr": reflect.ValueOf((*atomic.Uintptr)(nil)),
+		"Value":   reflect.ValueOf((*atomic.Value)(nil)),
 	}
 }
