@@ -1,7 +1,7 @@
 // Code generated by 'yaegi extract sync'. DO NOT EDIT.
 
-//go:build go1.17
-// +build go1.17
+//go:build go1.19
+// +build go1.19
 
 package sgolib
 
@@ -37,5 +37,9 @@ type _sync_Locker struct {
 	WUnlock func()
 }
 
-func (W _sync_Locker) Lock()   { W.WLock() }
-func (W _sync_Locker) Unlock() { W.WUnlock() }
+func (W _sync_Locker) Lock() {
+	W.WLock()
+}
+func (W _sync_Locker) Unlock() {
+	W.WUnlock()
+}
